+++ app/src/main/java/com/bonda/bonda/model/AccessTokenProvider.kt
@@ -1,4 +1,4 @@
-package com.bonda.bonda.util
+package com.bonda.bonda.model
 
 object AccessTokenProvider {
