+++ app/src/main/java/com/bonda/bonda/model/BookCategory.kt
@@ -22,6 +22,3 @@ enum class BookCategory(
         val BUSINESS_CATEGORIES: List<BookCategory> = entries.filter { it != ALL }
     }
 }
-
-fun String.toBookCategory(): BookCategory =
-    BookCategory.entries.find { it.code == this } ?: BookCategory.ALL
