+++ app/src/main/java/com/bonda/bonda/model/BookTheme.kt
@@ -19,6 +19,3 @@ enum class BookTheme(val code: String, val label: String) {
         val BUSINESS_THEMES: List<BookTheme> = entries.filter { it != ALL }
     }
 }
-
-fun String.toBookTheme(): BookTheme =
-    BookTheme.entries.find { it.code == this } ?: BookTheme.ALL
