+++ app/src/main/java/com/bonda/bonda/model/SortOrder.kt
@@ -6,6 +6,3 @@ enum class SortOrder(val code: String, val label: String) {
     RECENT("recent", "최신순"),
     POPULARITY("popularity", "인기순")
 }
-
-fun String.toSortOrder(): SortOrder =
-    SortOrder.entries.find { it.code == this } ?: SortOrder.RECENTLY_SAVED
