+++ app/src/main/java/com/bonda/bonda/ui/auth/SignInActivity.kt
@@ -27,6 +27,7 @@ import androidx.core.content.edit
 import androidx.core.net.toUri
 import com.bonda.bonda.R
 import com.bonda.bonda.databinding.ActivitySignInBinding
+import com.bonda.bonda.model.AccessTokenProvider
 import com.bonda.bonda.network.ApiClient
 import com.bonda.bonda.network.model.auth.LoginRequest
 
@@ -38,7 +39,6 @@ class SignInActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
-
         binding = ActivitySignInBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
@@ -52,9 +52,9 @@ class SignInActivity : AppCompatActivity() {
          * 카카오로 로그인 하기 버튼 로직
          */
         binding.signInButton.setOnClickListener {
-            var accessToken: String
-            var refreshToken: String
-            var isNewUser: Boolean
+            UserApiClient.instance.accessTokenInfo { tokenInfo, error ->
+                Log.d(TAG, "$tokenInfo, $error")
+            }
 
             /**
              * 카카오톡이 설치 되어 있는 경우
@@ -95,6 +95,7 @@ class SignInActivity : AppCompatActivity() {
              * 카카오톡이 설치되어있지 않은 경우 브라우저 로그인을 호출합니다
              */
             else {
+                Log.d(TAG, "로그인을 시도합니다")
                 UserApiClient.instance.loginWithKakaoAccount(this) { token, _ ->
                     if (token != null) {
                         lifecycleScope.launch {
