+++ app/src/main/java/com/bonda/bonda/ui/auth/SignUpActivity.kt
@@ -22,15 +22,15 @@ import com.bonda.bonda.databinding.ActivitySignUpBinding
 import com.bonda.bonda.network.ApiClient
 import com.bonda.bonda.ui.auth.onboarding.OnboardingActivity
 import com.bonda.bonda.ui.profile.ProfileImageSelectorView
-import com.bonda.bonda.util.AccessTokenProvider
+import com.bonda.bonda.model.AccessTokenProvider
 import com.bonda.bonda.util.PREFS_NAME
 import com.bonda.bonda.util.PREF_KEY_REFRESH_TOKEN
 import com.bonda.bonda.util.PREF_KEY_SIGNUP_REQUIRED
 import com.bonda.bonda.util.TAG
 import kotlinx.coroutines.launch
 import okhttp3.MediaType.Companion.toMediaType
+import okhttp3.MediaType.Companion.toMediaTypeOrNull
 import okhttp3.MultipartBody
-import okhttp3.RequestBody
 import okhttp3.RequestBody.Companion.toRequestBody
 
 class SignUpActivity : AppCompatActivity() {
@@ -86,7 +86,7 @@ class SignUpActivity : AppCompatActivity() {
                         val bytes =
                             this@SignUpActivity.contentResolver.openInputStream(uri)!!
                                 .use { it.readBytes() }
-                        val rb = RequestBody.create("image/*".toMediaType(), bytes)
+                        val rb = bytes.toRequestBody("image/*".toMediaTypeOrNull())
                         MultipartBody.Part.createFormData(
                             name = "profileImage",
                             filename = "profile.jpg",
