+++ app/src/main/java/com/bonda/bonda/ui/home/articles/ArticlesViewModel.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.bonda.bonda.AppEvents
+import com.bonda.bonda.model.AppEvents
 import com.bonda.bonda.model.ArticleCategory
 import com.bonda.bonda.network.ApiClient
 import com.bonda.bonda.util.TAG
