+++ app/src/main/res/layout/activity_onboarding.xml
@@ -15,7 +15,7 @@
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
-    <com.bonda.bonda.util.TabIndicatorView
+    <com.bonda.bonda.ui.components.TabIndicatorView
         android:id="@+id/page_indicator"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
